+++ Tests/JExtractSwiftTests/SpecializationTests.swift
@@ -263,7 +263,7 @@ struct SpecializationTests {
         @_cdecl("Java_com_example_swift_FishBox__00024observeTheFish__JJ")
         public func Java_com_example_swift_FishBox__00024observeTheFish__JJ(environment: UnsafeMutablePointer<JNIEnv?>!, thisClass: jclass, selfPointer: jlong, selfTypePointer: jlong) {
           assert(selfPointer != 0, "selfPointer memory address was null")
-          let selfPointerBits$ = Int(Int64(fromJNI: selfPointer, in: environment))
+          let selfPointerBits$ = Int(fromJNI: selfPointer, in: environment)
           let selfPointer$ = UnsafeMutablePointer<Box<Fish>>(bitPattern: selfPointerBits$)
           guard let selfPointer$ else {
             fatalError("selfPointer memory address was null in call to \\(#function)!")
@@ -276,12 +276,12 @@ struct SpecializationTests {
         @_cdecl("Java_com_example_swift_FishBox__00024count__JJ")
         public func Java_com_example_swift_FishBox__00024count__JJ(environment: UnsafeMutablePointer<JNIEnv?>!, thisClass: jclass, selfPointer: jlong, selfTypePointer: jlong) -> jlong {
           assert(selfPointer != 0, "selfPointer memory address was null")
-          let selfPointerBits$ = Int(Int64(fromJNI: selfPointer, in: environment))
+          let selfPointerBits$ = Int(fromJNI: selfPointer, in: environment)
           let selfPointer$ = UnsafeMutablePointer<Box<Fish>>(bitPattern: selfPointerBits$)
           guard let selfPointer$ else {
             fatalError("selfPointer memory address was null in call to \\(#function)!")
           }
-          return Int64(selfPointer$.pointee.count()).getJNILocalRefValue(in: environment)
+          return selfPointer$.pointee.count().getJNILocalRefValue(in: environment)
         }
         """,
       ],
